refactor: index package
