+++ internal/web/handler/index/new.go
@@ -12,27 +12,37 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <https://www.gnu.org/licenses/>
 
-package handler
+package index
 
 import (
-	"go.uber.org/fx"
+	"go.uber.org/zap"
 
-	"github.com/bangumi/server/internal/web/handler/character"
+	"github.com/bangumi/server/internal/cache"
+	"github.com/bangumi/server/internal/ctrl"
+	"github.com/bangumi/server/internal/domain"
 	"github.com/bangumi/server/internal/web/handler/common"
-	"github.com/bangumi/server/internal/web/handler/index"
-	"github.com/bangumi/server/internal/web/handler/person"
-	"github.com/bangumi/server/internal/web/handler/subject"
-	"github.com/bangumi/server/internal/web/handler/user"
 )
 
-var Module = fx.Module("handler",
-	fx.Provide(
-		New,
-		common.New,
-		user.New,
-		person.New,
-		subject.New,
-		character.New,
-		index.New,
-	),
-)
+type Handler struct {
+	common.Common
+	ctrl  ctrl.Ctrl
+	cache cache.RedisCache
+	i     domain.IndexRepo
+	log   *zap.Logger
+}
+
+func New(
+	common common.Common,
+	index domain.IndexRepo,
+	log *zap.Logger,
+	cache cache.RedisCache,
+	ctrl ctrl.Ctrl,
+) Handler {
+	return Handler{
+		Common: common,
+		ctrl:   ctrl,
+		cache:  cache,
+		log:    log.Named("web.handler"),
+		i:      index,
+	}
+}
